pyproject.toml
[project]
name = "stress-indicator"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.52.2,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "google-genai (>=1.56.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
